Extract Kamal::EnvFile
Cleaning up the Utils junk drawer.
